Refactor variable declaration handling in typechecker and move checkImplStmt to impl.go
